Backfill a few missing columns in equities.csv
Wrote a quick script to backfill any missing columns in equities.csv.
The script will keep any existing values, with the exception of market_cap.
market_cap tended to be stale in several entries I looked at.
The market_cap recomputation matches the update logic:
https://github.com/JerBouma/FinanceDatabase/blob/main/.github/workflows/database_update.yml#L61-L77

The end goal of the script is to add missing any ISIN
as it is keyed off of those.

This PR contains only a small amount of changes to confirm the
expected behavior and the desirability of running this further.